// src/api-client/models/inventory-facets-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spira API Documentation
 * Spira API
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface InventoryFacetsResponseDto
 */
export interface InventoryFacetsResponseDto {
    /**
     * How many lots matched the status and branch.
     * @type {number}
     * @memberof InventoryFacetsResponseDto
     */
    'total': number;
    /**
     * Count per category. Every member is present, including zeros, so the chips do not appear and disappear.
     * @type {{ [key: string]: number; }}
     * @memberof InventoryFacetsResponseDto
     */
    'byCategory': { [key: string]: number; };
    /**
     * Count per urgency band, derived from the expiries.
     * @type {{ [key: string]: number; }}
     * @memberof InventoryFacetsResponseDto
     */
    'byUrgency': { [key: string]: number; };
}