// src/api-client/models/expiry-kind.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spira API Documentation
 * Spira API
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * Which kind of date `expiresAt` was, as snapshotted at handover. Best-before is a quality date; use-by is a safety one.
 * @export
 * @enum {string}
 */

export const ExpiryKind = {
    BestBefore: 'BEST_BEFORE',
    UseBy: 'USE_BY'
} as const;

export type ExpiryKind = typeof ExpiryKind[keyof typeof ExpiryKind];